.github/scripts/requirements.txt
requests
python-jose
urllib3
python-keycloak
boto3
